Some changes in releasing

+++ entrypoint.sh
@@ -84,10 +84,22 @@ if ! make $make_opts -j"$(nproc --all)"; then
 fi
 msg "Packaging the kernel..."
 zip_filename="${NAME:-$repo_name}-${tag}-${date}.zip"
-git clone --depth 1 https://"$zipper".git $zipper_path || exit 127
-cp out/arch/"$arch"/boot/"$image" "$zipper_path"/"$image"
-cd $zipper_path || exit 127
-rm -rf .git
-zip -r9 "$zip_filename" . || exit 127
-outfile "$zipper_path"/"$zip_filename"
-cd "$workdir" || exit 127
+if [[ $zipper = "nozipper" ]]; then
+    msg "No zip template provided, releasing the kernel image instead"
+    outfile out/arch/"$arch"/boot/"$image"
+    exit 0
+else
+    if ! git clone --depth 1 https://"$zipper".git $zipper_path; then
+        msg "Error cloning zip template, releasing the kernel image instead"
+        outfile out/arch/"$arch"/boot/"$image"
+        cd "$workdir" || exit 127
+        exit 0
+    fi    
+    cp out/arch/"$arch"/boot/"$image" "$zipper_path"/"$image"
+    cd $zipper_path || exit 127
+    rm -rf .git
+    zip -r9 "$zip_filename" . || exit 127
+    outfile "$zipper_path"/"$zip_filename"
+    cd "$workdir" || exit 127
+    exit 0
+fi
